requirements.txt
boto3==1.34.0
moviepy==2.1.2
numpy>=1.21.0
requests==2.31.0
python-dotenv==1.0.0
imageio[ffmpeg]